feat: dstew 2268 prior authority decision endpoint (#482)
* tests

* add MakePriorAuthorityDecisionRequest

* add tests for test coverage

* add priorAuthorityType to event

* added PA make decision to projection

* add test for test coverage

* added PA make decision to projection

* added handler around PA make decision

* add fields to event

* added exert fee information to pa make decision

* add test for expert

* jacoco

* jacoco

* added exert fee information to pa make decision

* added exert fee information to tests for
pa make decision

* replaced individual expert fields with the object in the payload

* added disbursement and apportionment support and tests

* added disbursement and apportionment support and tests

* update test

* added integration test stub

* added integration tests to cover PA make decision

* spotless

* fix pipeline

* fix merge issues

* work lists

* update read side

* tidy up

* address review comments

* revert status changes

* fix failing test

* fix faililng test

* testing

* update to use OID caseworker

* change financial values to bigDecimal

* change submissionId to priorAuthorityId

* add decision package

* add DECIDED to PA lifecycle

* remove status from PriorAuthorityState

* rename PriorAuthorityDecisionRecordedEvent to PriorAuthorityDecisionMadeEvent

* add lombok.builder to priorauthorityresult

* remove decisionJustification from event data storage

* add decisionDetails to priorauthorityresult

* jacoco

* changes following review

* changes after review

* fix tests

* fix flyway issue

* change switch ordering

* fixing issues

* fix

---------

+++ data-access-service-axon/src/test/java/uk/gov/justice/laa/dstew/access/controller/application/SavePriorAuthorityDraftCommandMapperTest.java
@@ -198,14 +198,14 @@ class SavePriorAuthorityDraftCommandMapperTest {
                 .expertCosts(
                     ExpertCosts.builder()
                         .billingType(BillingType.HOURLY)
-                        .hourlyRate(300.0)
+                        .hourlyRate(BigDecimal.valueOf(300.0))
                         .timeRequested(TimeRequested.builder().hours(2).minutes(30).build())
-                        .totalAmount(900.0)
+                        .totalAmount(BigDecimal.valueOf(900.0))
                         .costsSharedWithOtherParties(true)
                         .apportionment(
                             Apportionment.builder()
                                 .partiesSharingCosts(2)
-                                .clientShareAmount(450.0)
+                                .clientShareAmount(BigDecimal.valueOf(450.0))
                                 .build())
                         .build())
                 .build())
@@ -230,7 +230,7 @@ class SavePriorAuthorityDraftCommandMapperTest {
         .disbursementDetails(
             DisbursementDetails.builder()
                 .disbursementPurpose("Interpreter")
-                .disbursementAmount(150.25)
+                .disbursementAmount(BigDecimal.valueOf(150.25))
                 .build())
         .build();
   }
